refactor: move specifications to folder
