+++ Source/Application/Kysect.Shreks.Application.Handlers/Google/GetCoursePointsBySubjectCourseHandler.cs
@@ -1,7 +1,8 @@
-﻿using AutoMapper;
+using AutoMapper;
 using Kysect.Shreks.Application.Dto.Study;
 using Kysect.Shreks.Application.Dto.Tables;
 using Kysect.Shreks.Application.Dto.Users;
+using Kysect.Shreks.Application.Specifications;
 using Kysect.Shreks.Core.DeadlinePolicies;
 using Kysect.Shreks.Core.Study;
 using Kysect.Shreks.Core.Submissions;
